+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
   The first provides a wrapper for `qnethernet_hal_millis()`. The second wraps
   the ARM DWT_CYCCNT cycle counter.
 * Added printing chrono clock times to network-up in the main program.
+* Added `qindesign::network::util::elapsedTime<Clock>`, a class similar to
+  `elapsedMillis`, but uses a `std::chrono` Clock.
 
 ### Changed
 * Improved _Ping_ example by always sending, even if a send fails. It was